Refactor code to handle invalid product ID in deleteProduct API endpoint

+++ backend/controllers/product.controller.js
@@ -58,6 +58,12 @@ export const deleteProduct = async (req, res) => {
   const { id } = req.params;
   // console.log('id -', id);
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product ID" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res
@@ -65,6 +71,6 @@ export const deleteProduct = async (req, res) => {
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.error("Error in deleting product -", error.message);
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Internal Server Error" });
   }
 };
